Only drop the wfs table if it exists. Fixes #1110

+++ db/migrate/20180510050959_drop_wfs_rails_workflows.rb
@@ -1,5 +1,10 @@
 class DropWfsRailsWorkflows < ActiveRecord::Migration[5.1]
   def change
+    # This migration was created by a gem that was only loaded in development and
+    # test modes. So, dropping the table in production/staging would fail.
+    # Additionally, the drop_table would fail for anyone who has newly created
+    # their database and had never had the wfs_rails gem installed
+    return unless ActiveRecord::Base.connection.table_exists? 'wfs_rails_workflows'
     drop_table :wfs_rails_workflows
   end
 end
